Update docs and remove separate wasm CI step
Integrated into the appropriate existing steps.
