+++ ci/image/Dockerfile
@@ -6,16 +6,16 @@ RUN curl -sS https://dl.yarnpkg.com/debian/pubkey.gpg | apt-key add - && \
 RUN curl -sL https://deb.nodesource.com/setup_12.x | bash -
 RUN apt-get install -y nodejs chromium yarn
 
-RUN git config --global color.ui always
+COPY ./ci/image/gitignore /etc/git/ignore
+RUN git config --system color.ui always
+# Need to set set this explicitly for the system since github uses HOME=/home/github.
+RUN git config --system core.excludesfile /etc/git/ignore
 
 ENV GOPATH=/root/gopath
 ENV PATH=$GOPATH/bin:$PATH
 ENV GOFLAGS="-mod=readonly"
 ENV PAGER=cat
 ENV CI=true
-
-RUN mkdir -p ~/.config/git
-COPY ./ci/image/gitignore ~/.config/git/ignore
 
 # Cache go modules, build cache and yarn cache.
 COPY . /tmp/websocket
