+++ ci/lint.ts
@@ -1,8 +1,8 @@
 #!/usr/bin/env -S npx ts-node -P ci/tsconfig.json
 
-import { exec, main } from "./lib"
+import { exec, main, wasmEnv } from "./lib"
 
-if (process.argv[1] === __filename) {
+if (require.main === module) {
   main(lint)
 }
 
@@ -13,6 +13,15 @@ export async function lint(ctx: Promise<unknown>) {
       exec(ctx, "git ls-files '*.ts' | xargs npx eslint --max-warnings 0 --fix", {
         cwd: "ci",
       }),
+      wasmLint(ctx),
     ],
   )
+}
+
+
+async function wasmLint(ctx: Promise<unknown>) {
+  await exec(ctx, "go install golang.org/x/lint/golint")
+  await exec(ctx, "golint -set_exit_status ./...", {
+    env: wasmEnv,
+  })
 }
