+++ package.json
@@ -7,6 +7,7 @@
 		"@typescript-eslint/eslint-plugin": "^2.3.2",
 		"@typescript-eslint/parser": "^2.3.2",
 		"await-timeout": "^0.6.0",
+		"axios": "^0.19.0",
 		"child-process-promise": "^2.2.1",
 		"eslint": "^6.5.1",
 		"eslint-config-prettier": "^6.3.0",
@@ -20,7 +21,6 @@
 		"fmt": "./ci/fmt.ts",
 		"lint": "./ci/lint.ts",
 		"test": "./ci/test.ts",
-		"wasm": "./ci/wasm.ts",
 		"ci": "./ci/all.ts",
 		"push-ci-image": "ci/image/push.ts"
 	}
